requirements.txt
paddlepaddle
paddleocr
fastapi
uvicorn
opencv-python
python-multipart
Pillow
